LayoutTests/http/tests/webAPIStatistics/canvas-read-and-write-data-collection-expected.txt

Tests for canvas read and write data collection in ResourceLoadStatistics plist by rendering and reading text on the canvas and dumping the entire resource load statistics map.
Resource load statistics:

Registrable domain: 127.0.0.1
    hadUserInteraction: No
    mostRecentUserInteraction: -1
    grandfathered: No
    IsScheduledForAllButCookieDataRemoval: No
    isPrevalentResource: No
    isVeryPrevalentResource: No
    dataRecordsRemoved: 0
